externalisation des différentes fonctions
